Added GitLab CI and Kia Linter
Added the CI to run the test suite. Added the kia linter
which makes sure markdown files are formatted correctly.
Included are sample directories which the app will be
built upon. Each directory has an ABOUTDIR.md file
which is required by the kia linter. Updated the package.json
with a script to run the kia linter.

-	new file:   .gitlab-ci.yml
-	new file:   app/ABOUTDIR.md
-	new file:   app/components/ABOUTDIR.md
-	new file:   app/config/ABOUTDIR.md
-	new file:   app/images/ABOUTDIR.md
-	new file:   app/layouts/ABOUTDIR.md
-	new file:   app/lib/ABOUTDIR.md
-	new file:   app/routes/ABOUTDIR.md
-	new file:   kiaLinter.js
-	modified:   package.json